bronze/requirements.txt
Flask
requests
zstandard
google-cloud-pubsub